refactor: Rename Rust API module from 'simple' to 'hurl_parser' and consolidate tests
- Renamed rust/src/api/simple.rs to hurl_parser.rs for clearer intent
- Updated all FFI binding references in Dart generated code
- Deleted scattered test files (parser_test.dart, enhanced_parser_test.dart, advanced_features_test.dart)
- Created single consolidated hurl_parser_test.dart with 13 tests in 7 logical groups
- Removed apidash_core hurl_io_test.dart (FFI limitations make it impractical)
- Updated comment style in hurl_io.dart for consistency
- Regenerated Flutter Rust Bridge bindings with correct content hash
