+++ packages/hurl/lib/hurl.dart
@@ -1,4 +1,4 @@
 library;
 
-export 'src/rust/api/simple.dart';
+export 'src/rust/api/hurl_parser.dart';
 export 'src/rust/frb_generated.dart' show RustLib;
