+++ packages/hurl/lib/src/rust/frb_generated.io.dart
@@ -3,7 +3,7 @@
 
 // ignore_for_file: unused_import, unused_element, unnecessary_import, duplicate_ignore, invalid_use_of_internal_member, annotate_overrides, non_constant_identifier_names, curly_braces_in_flow_control_structures, prefer_const_literals_to_create_immutables, unused_field
 
-import 'api/simple.dart';
+import 'api/hurl_parser.dart';
 import 'dart:async';
 import 'dart:convert';
 import 'dart:ffi' as ffi;
