+++ packages/hurl/lib/src/rust/frb_generated.web.dart
@@ -6,7 +6,7 @@
 // Static analysis wrongly picks the IO variant, thus ignore this
 // ignore_for_file: argument_type_not_assignable
 
-import 'api/simple.dart';
+import 'api/hurl_parser.dart';
 import 'dart:async';
 import 'dart:convert';
 import 'frb_generated.dart';
