+++ packages/hurl/rust/src/api/mod.rs
@@ -1,1 +1,1 @@
-pub mod simple;
+pub mod hurl_parser;
